requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.0.1
python-dotenv==0.19.1
flask-jwt-extended==4.3.1
bcrypt==3.2.0
python-slugify==5.0.2
Werkzeug==2.0.1
pillow>=10.0.0